Refactor blog category imports to use centralized service module
