+++ src/app/blogs/categories/[slug]/page.tsx
@@ -2,9 +2,8 @@ import { Metadata } from "next";
 import { BlogCategory } from "@prisma/client";
 import { getBlogsByCategory } from "@/lib/services/blog";
 import { BlogWithDetailsType } from "@/components/blog/types/blogTypes";
-import { getBlogCategories } from "@/components/blog/services/blogCategoryCrud"; // Updated import path
+import { getBlogCategories, getBlogCategoryByName } from "@/lib/services/blog";
 import BlogCard from "@/components/blog/blog-components/BlogCard";
-import { getBlogCategoryByName } from "@/components/blog/services/blogCategoryCrud";
 
 export async function generateMetadata({
   params,
